Send Message to Server
Rearranged the formatting so that the routing logic is in a function

Changed SendMessage() so that it takes the current position and returns the value of the next position as told by the server

Debugged Error: assert failed: heap_caps_free heap_caps_base.c:63 (heap != NULL && "free() target pointer is outside heap areas")

This error shows that there was an issue w dynamic memory allocation within the program. Indicates that it was attempting to free memory that was either not allocated or already freed.